[S09P12B207-350] feat: instance.ts 수정

+++ frontend-web/src/apis/instance.ts
@@ -15,7 +15,6 @@ instance.interceptors.response.use(
   (response: AxiosResponse) => {
     // 헤더를 뜯어서 accessToken을 저장한다
     const accessToken = response.headers['x-access-token'];
-
     if (accessToken !== null && accessToken !== undefined) {
       localStorage.setItem('accessToken', accessToken);
     }
@@ -23,10 +22,11 @@ instance.interceptors.response.use(
     return response;
   },
   (error: AxiosError) => {
-    console.log(error);
-    // 로그아웃 시키고 로그인 PAGE로 리다이렉트
-    localStorage.clear();
-    window.location.href = PATH.LOGIN;
+    if (error.response?.status === 401) {
+      // 로그아웃 시키고 로그인 PAGE로 리다이렉트
+      localStorage.clear();
+      window.location.href = PATH.LOGIN;
+    }
   }
 );
